# crates/userserver/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BSCP API",
    "version": "1.0.0",
    "description": "Beorn's Stupid Chat Protocol — user server API. This document is only served by debug builds."
  },
...
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "bscp_session" },
      "sessionHeader": { "type": "apiKey", "in": "header", "name": "X-Session-Token" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "errors": { "type": "array", "items": { "type": "string" } },
          "code": { "type": "integer" }
        }
      },
      "Profile": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "example": "alice@localhost:5000" },
          "display_name": { "type": "string" },
          "profile_pic": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["online", "offline", "away", "dnd"] },
          "is_admin": { "type": "boolean" },
          "is_primary_admin": { "type": "boolean" },
          "is_2fa_enabled": { "type": "boolean" },
          "bio": { "type": "string", "nullable": true },
          "storage_limit_mb": { "type": "integer" }
        }
      },
...
      "Chat": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "display_name": { "type": "string" },
          "profile_pic": { "type": "string", "nullable": true },
          "status": { "type": "string" },
          "unread_count": { "type": "integer" },
          "last_message_text": { "type": "string", "nullable": true },
          "last_message_sender": { "type": "string", "nullable": true }
        }
      },
      "Invite": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "code": { "type": "string" },
          "status": { "type": "string" },
          "created_at": { "type": "number", "nullable": true },
          "expires_at": { "type": "number", "nullable": true },
          "used_by": { "type": "string", "nullable": true }
        }
      },
      "Webhook": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "url": { "type": "string" },
          "profile_pic": { "type": "string", "nullable": true },
          "created_at": { "type": "number" },
          "last_used": { "type": "number", "nullable": true }
        }
      }
    }
  },
  "paths": {
    "/api/auth/setup": {
      "get": {
        "tags": ["auth"],
        "summary": "Whether first-time setup is required",
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "needs_setup": { "type": "boolean" } } } } } } }
      },
      "post": {
        "tags": ["auth"],
        "summary": "Create the initial admin account",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["username", "password", "password_confirm"], "properties": { "username": { "type": "string" }, "email": { "type": "string", "nullable": true }, "password": { "type": "string" }, "password_confirm": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Authenticate with username & password",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["user", "password"], "properties": { "user": { "type": "string" }, "password": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Login result", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "requires_2fa": { "type": "boolean" }, "error": { "type": "string" }, "session_token": { "type": "string" } } } } } } }
      }
    },
...
    "/api/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new account with an invite code",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["username", "password", "password_confirm", "invite_code"], "properties": { "username": { "type": "string" }, "password": { "type": "string" }, "password_confirm": { "type": "string" }, "invite_code": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
...
    "/api/users/me": {
      "get": { "tags": ["users"], "summary": "Current user's profile", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } } } }, "401": { "description": "Unauthenticated" } } },
      "patch": { "tags": ["users"], "summary": "Update display name / bio", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "display_name": { "type": "string" }, "bio": { "type": "string", "nullable": true } } } } } }, "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } } } } } }
    },
    "/api/users/": {
      "get": { "tags": ["users"], "summary": "List all users (admin)", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Profile" } } } } }, "403": { "description": "Not an admin" } } }
    },
...
    "/api/chats/": {
      "get": { "tags": ["chats"], "summary": "List conversations", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Chat" } } } } } } }
    },
...
    "/api/upload/": {
      "post": { "tags": ["uploads"], "summary": "Upload a file, returns a markdown embed", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" } } } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "type": "object", "properties": { "url": { "type": "string" }, "mimetype": { "type": "string" }, "markdown": { "type": "string" } } } } } }, "413": { "description": "Over quota" } } }
    },
    "/api/upload/{upload_id}": {
      "delete": { "tags": ["uploads"], "summary": "Delete one of your uploads", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "parameters": [{ "name": "upload_id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "204": { "description": "Deleted" } } }
    },
    "/api/upload/user/list": {
      "get": { "tags": ["uploads"], "summary": "List your uploads and quota usage", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "uploads": { "type": "array", "items": { "type": "object" } }, "total_size_bytes": { "type": "integer" }, "limit_bytes": { "type": "integer" } } } } } } } }
    },
    "/api/invites/": {
      "get": { "tags": ["invites"], "summary": "List invite codes (admin)", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Invite" } } } } } } }
    },
...
    "/api/user/webhooks": {
      "get": { "tags": ["webhooks"], "summary": "List your webhooks", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Webhook" } } } } } } },
      "post": { "tags": ["webhooks"], "summary": "Create a webhook (name is immutable)", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "avatar_url": { "type": "string", "nullable": true } } } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Webhook" } } } } } }
    },
...
    "/webhooks/{webhook_id}/{webhook_token}": {
      "post": { "tags": ["webhooks"], "summary": "Incoming webhook endpoint (external callers)", "parameters": [{ "name": "webhook_id", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "webhook_token", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["content"], "properties": { "content": { "type": "string" }, "username": { "type": "string" }, "avatar_url": { "type": "string" } } } } } }, "responses": { "201": { "description": "Created" }, "404": { "description": "Invalid webhook" } } }
    },
    "/media/proxy": {
      "get": { "summary": "Cached media proxy", "parameters": [{ "name": "url", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "The media bytes" }, "400": { "description": "Missing URL" }, "502": { "description": "Fetch failed" } } }
    },
    "/federation/receive": {
      "post": { "tags": ["federation"], "summary": "Receive a federated message", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "id": { "type": "string" }, "sender": { "type": "string" }, "receiver": { "type": "string" }, "text": { "type": "string" }, "validationKey": { "type": "string" } } } } } }, "responses": { "200": { "description": "Stored" }, "401": { "description": "Validation failed" } } }
    },
    "/federation/validate": {
      "get": { "tags": ["federation"], "summary": "Confirm a message originated here", "parameters": [{ "name": "messageId", "in": "query", "schema": { "type": "string" } }, { "name": "validationKey", "in": "query", "schema": { "type": "string" } }, { "name": "sender", "in": "query", "schema": { "type": "string" } }, { "name": "receiver", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "valid": { "type": "boolean" } } } } } } } }
    },
...
  }
}